+++ apps/myaccount/src/components/account-recovery/options/index.ts
@@ -18,4 +18,4 @@
 
 export * from "./email-recovery";
 export * from "./security-questions-recovery";
-export * from "./sms-otp-recovery";
+export * from "./sms-recovery/sms-recovery";
