+++ apps/myaccount/src/components/account-recovery/options/sms-recovery/sms-recovery.scss
@@ -1,0 +1,16 @@
+form > div.oxygen-text-field > div {
+    width: 200px !important;
+}
+
+form > p{
+    margin-top: 10px !important;
+    font-size: 12px;
+}
+
+div.description {
+    margin-top: 5px;
+}
+
+form > div.inline.fields > button {
+    margin-top: 8px;
+}
