+++ modules/i18n/src/models/namespaces/myaccount-ns.ts
@@ -79,7 +79,7 @@ export interface MyAccountNS {
             }
         }
         accountRecovery: {
-            SMSOTPRecovery: {
+            SMSRecovery: {
                 descriptions: {
                     add: string;
                     update: string;
